Extract local open targets package
