+++ packages/local-open-targets/src/errors.ts
@@ -1,0 +1,22 @@
+export type WorkspaceOpenTargetErrorCode =
+  | "path_not_found"
+  | "path_not_openable"
+  | "remote_mapping_missing"
+  | "remote_target_unsupported"
+  | "target_unavailable"
+  | "unsupported_platform";
+
+export interface WorkspaceOpenTargetErrorOptions {
+  code: WorkspaceOpenTargetErrorCode;
+  message: string;
+}
+
+export class WorkspaceOpenTargetError extends Error {
+  readonly code: WorkspaceOpenTargetErrorCode;
+
+  constructor(options: WorkspaceOpenTargetErrorOptions) {
+    super(options.message);
+    this.name = "WorkspaceOpenTargetError";
+    this.code = options.code;
+  }
+}
