+++ packages/local-open-targets/src/index.ts
@@ -7,12 +7,48 @@ import { z } from "zod";
 import {
   WORKSPACE_OPEN_TARGET_ICON_DATA_URL_MAX_LENGTH,
   type WorkspaceOpenTarget,
-  type WorkspaceOpenTargetCapabilities,
   type WorkspaceOpenTargetIcon,
   type WorkspaceOpenTargetId,
-  type WorkspaceOpenTargetKind,
 } from "@bb/host-daemon-contract";
 import { sanitizeInheritedChildProcessEnv } from "@bb/process-utils";
+import {
+  BASIC_FILE_OPEN_CAPABILITIES,
+  FILE_MANAGER_OPEN_CAPABILITIES,
+  TERMINAL_OPEN_CAPABILITIES,
+} from "./capabilities.js";
+import { WorkspaceOpenTargetError } from "./errors.js";
+import { LAUNCH_ADAPTERS } from "./macos-launch-adapters.js";
+import {
+  buildLocalTerminalShellArgs,
+  buildRemoteTerminalSshArgs,
+} from "./terminal.js";
+import type {
+  BuildMacRemoteSshOpenArgs,
+  BuildMacTerminalOpenArgs,
+  DiscoveredMacApplication,
+  ExecFileResult,
+  ExecFileInvocation,
+  ExistingPath,
+  LaunchAdapter,
+  LinuxDesktopApplication,
+  ListWorkspaceOpenTargetsOptions,
+  MacRemoteSshOpenCommandAdapter,
+  OpenPathInTargetArgs,
+  PlatformOpenInvocationArgs,
+  PlatformRemoteSshOpenInvocationArgs,
+  ResolveMacOpenInvocationArgs,
+  ResolveMacRemoteSshOpenInvocationArgs,
+  ResolveMacTargetIconArgs,
+  ResolveTargetOpenPathArgs,
+  WorkspaceOpenTargetRuntime,
+} from "./types.js";
+
+export { WorkspaceOpenTargetError } from "./errors.js";
+export type {
+  ListWorkspaceOpenTargetsOptions,
+  OpenPathInTargetArgs,
+  WorkspaceOpenTargetRuntime,
+} from "./types.js";
 
 const execFileAsync = promisify(execFile);
 const DESKTOP_APP_TARGET_ID_PREFIX = "desktop-app:";
@@ -46,1004 +82,6 @@ const macApplicationsForFileResultSchema = z.array(
     })
     .strict(),
 );
-
-export type WorkspaceOpenTargetErrorCode =
-  | "path_not_found"
-  | "path_not_openable"
-  | "remote_mapping_missing"
-  | "remote_target_unsupported"
-  | "target_unavailable"
-  | "unsupported_platform";
-
-export interface WorkspaceOpenTargetErrorOptions {
-  code: WorkspaceOpenTargetErrorCode;
-  message: string;
-}
-
-export class WorkspaceOpenTargetError extends Error {
-  readonly code: WorkspaceOpenTargetErrorCode;
-
-  constructor(options: WorkspaceOpenTargetErrorOptions) {
-    super(options.message);
-    this.name = "WorkspaceOpenTargetError";
-    this.code = options.code;
-  }
-}
-
-export interface OpenPathInTargetArgs {
-  columnNumber: number | null;
-  context: OpenPathInTargetContext;
-  lineNumber: number | null;
-  path: string;
-  targetId: WorkspaceOpenTargetId;
-}
-
-export interface ListWorkspaceOpenTargetsOptions {
-  path?: string;
-}
-
-interface ExecFileResult {
-  stdout: string;
-}
-
-type ExecFileHandler = (
-  file: string,
-  args: string[],
-) => Promise<ExecFileResult>;
-
-export interface WorkspaceOpenTargetRuntime {
-  applicationDirectories: string[];
-  desktopFileDirectories?: string[];
-  env?: NodeJS.ProcessEnv;
-  execFile: ExecFileHandler;
-  platform: NodeJS.Platform;
-}
-
-interface MacDefaultLaunchAdapter {
-  openMode: "default-app";
-}
-
-interface MacApplicationLaunchAdapter {
-  appName: string;
-  bundleIds: string[];
-  builtIn: boolean;
-  fileOpenCommand?: MacFileOpenCommandAdapter;
-  lineOpenCommand?: MacLineOpenCommandAdapter;
-  localTerminalOpenCommand?: MacLocalTerminalOpenCommandAdapter;
-  openMode: "application";
-  pathOpenCommand?: MacPathOpenCommandAdapter;
-  remoteSshOpenCommand?: MacRemoteSshOpenCommandAdapter;
-}
-
-type MacLaunchAdapter = MacApplicationLaunchAdapter | MacDefaultLaunchAdapter;
-
-interface LaunchAdapter {
-  capabilities: WorkspaceOpenTargetCapabilities;
-  fileOpenBehavior: "direct" | "containing-directory";
-  icon: WorkspaceOpenTargetIcon;
-  id: WorkspaceOpenTargetId;
-  kind: WorkspaceOpenTargetKind;
-  label: string;
-  macos: MacLaunchAdapter;
-}
-
-interface ExecFileInvocation {
-  args: string[];
-  file: string;
-}
-
-interface BuildMacLineOpenArgs {
-  columnNumber: number | null;
-  lineNumber: number;
-  path: string;
-}
-
-interface BuildMacTerminalOpenArgs {
-  columnNumber: number | null;
-  lineNumber: number | null;
-  path: string;
-}
-
-interface BuildMacRemoteSshOpenArgs {
-  columnNumber: number | null;
-  lineNumber: number | null;
-  path: string;
-  sshAuthority: string;
-}
-
-interface MacLineOpenCommandAdapter {
-  executable: string;
-  supportsColumn: boolean;
-  toArgs: (args: BuildMacLineOpenArgs) => string[];
-}
-
-interface MacPathOpenCommandAdapter {
-  executable: string;
-  toArgs: (path: string) => string[];
-}
-
-interface MacFileOpenCommandAdapter {
-  executable: string;
-  toArgs: (path: string) => string[];
-}
-
-interface MacLocalTerminalOpenCommandAdapter {
-  executable: string;
-  toArgs: (args: BuildMacTerminalOpenArgs) => string[];
-}
-
-interface MacRemoteSshOpenCommandAdapter {
-  capabilities: WorkspaceOpenTargetCapabilities;
-  executable: string;
-  requiredExecutables?: string[];
-  toArgs: (args: BuildMacRemoteSshOpenArgs) => string[];
-}
-
-interface PlatformOpenInvocationArgs {
-  columnNumber: number | null;
-  existingPath: ExistingPath;
-  lineNumber: number | null;
-}
-
-interface PlatformRemoteSshOpenInvocationArgs {
-  columnNumber: number | null;
-  lineNumber: number | null;
-  path: string;
-  sshAuthority: string;
-}
-
-interface ResolveMacOpenInvocationArgs {
-  columnNumber: number | null;
-  definition: LaunchAdapter;
-  existingPath: ExistingPath;
-  lineNumber: number | null;
-}
-
-interface ResolveMacRemoteSshOpenInvocationArgs {
-  columnNumber: number | null;
-  definition: LaunchAdapter;
-  lineNumber: number | null;
-  path: string;
-  sshAuthority: string;
-}
-
-interface ResolveTargetOpenPathArgs {
-  definition: LaunchAdapter;
-  existingPath: ExistingPath;
-}
-
-interface ResolveMacTargetIconArgs {
-  appPath: string | null;
-  definition: LaunchAdapter;
-  runtime: WorkspaceOpenTargetRuntime;
-}
-
-interface DiscoveredMacApplication {
-  appPath: string;
-  bundleId: string | null;
-  label: string;
-}
-
-interface LinuxDesktopApplication {
-  desktopFileId: string;
-  desktopFilePath: string;
-  exec: string;
-  label: string;
-}
-
-type OpenPathInTargetContext =
-  | { kind: "local" }
-  | {
-      kind: "remote-ssh";
-      serverOrigin: string;
-      hostId: string;
-      sshAuthority: string;
-    };
-
-function formatPathWithLineNumber(args: BuildMacLineOpenArgs): string {
-  return args.columnNumber === null
-    ? `${args.path}:${args.lineNumber}`
-    : `${args.path}:${args.lineNumber}:${args.columnNumber}`;
-}
-
-function formatJetBrainsLineOpenArgs(args: BuildMacLineOpenArgs): string[] {
-  return ["--line", String(args.lineNumber), args.path];
-}
-
-function formatZedRemoteSshUri(args: BuildMacRemoteSshOpenArgs): string {
-  const absolutePath = args.path.startsWith("/") ? args.path : `/${args.path}`;
-  const encodedPath = absolutePath.split("/").map(encodeURIComponent).join("/");
-  const uri = `ssh://${args.sshAuthority}${encodedPath}`;
-  if (args.lineNumber === null) {
-    return uri;
-  }
-  return args.columnNumber === null
-    ? `${uri}:${args.lineNumber}`
-    : `${uri}:${args.lineNumber}:${args.columnNumber}`;
-}
-
-function quoteShellArg(value: string): string {
-  if (value === "") {
-    return "''";
-  }
-  return `'${value.replaceAll("'", "'\\''")}'`;
-}
-
-function escapeAppleScriptString(value: string): string {
-  return value.replaceAll("\\", "\\\\").replaceAll('"', '\\"');
-}
-
-function buildTerminalShellScript(args: BuildMacTerminalOpenArgs): string {
-  const line = args.lineNumber === null ? "" : String(args.lineNumber);
-  const column = args.columnNumber === null ? "" : String(args.columnNumber);
-  const editorScript = [
-    "editor=''",
-    'if [ -n "${VISUAL:-}" ]; then editor=$VISUAL; elif [ -n "${EDITOR:-}" ]; then editor=$EDITOR; elif command -v nvim >/dev/null 2>&1; then editor=nvim; elif command -v vim >/dev/null 2>&1; then editor=vim; elif command -v vi >/dev/null 2>&1; then editor=vi; elif command -v nano >/dev/null 2>&1; then editor=nano; elif command -v less >/dev/null 2>&1; then editor=less; fi',
-    'if [ -n "$editor" ]; then editor_name=$(basename "$editor"); case "$editor_name" in nvim|vim|vi) if [ -n "$line" ] && [ -n "$column" ]; then "$editor" "+call cursor($line,$column)" "$file"; elif [ -n "$line" ]; then "$editor" "+$line" "$file"; else "$editor" "$file"; fi ;; nano) if [ -n "$line" ] && [ -n "$column" ]; then "$editor" "+$line,$column" "$file"; elif [ -n "$line" ]; then "$editor" "+$line" "$file"; else "$editor" "$file"; fi ;; less) if [ -n "$line" ]; then "$editor" "+${line}g" "$file"; else "$editor" "$file"; fi ;; *) "$editor" "$file" ;; esac; else printf "%s\\n" "No terminal editor found for $target"; fi',
-  ].join("; ");
-
-  return [
-    `target=${quoteShellArg(args.path)}`,
-    `line=${quoteShellArg(line)}`,
-    `column=${quoteShellArg(column)}`,
-    'if [ -d "$target" ]; then cd "$target" || exit; exec "${SHELL:-/bin/sh}"; fi',
-    'if [ -f "$target" ]; then dir=$(dirname "$target") || exit; file=$(basename "$target") || exit; cd "$dir" || exit',
-    editorScript,
-    'exec "${SHELL:-/bin/sh}"',
-    "fi",
-    'printf "%s\\n" "Path is not a file or directory: $target"',
-    'exec "${SHELL:-/bin/sh}"',
-  ].join("; ");
-}
-
-function buildLocalTerminalShellArgs(args: BuildMacTerminalOpenArgs): string[] {
-  return ["/bin/sh", "-lc", buildTerminalShellScript(args)];
-}
-
-function buildRemoteTerminalSshArgs(args: BuildMacRemoteSshOpenArgs): string[] {
-  return ["ssh", "-t", "--", args.sshAuthority, buildTerminalShellScript(args)];
-}
-
-function buildShellCommand(args: string[]): string {
-  return args.map(quoteShellArg).join(" ");
-}
-
-function buildMacTerminalLocalOpenArgs(
-  args: BuildMacTerminalOpenArgs & { appName: "Terminal" | "iTerm" },
-): string[] {
-  const command = escapeAppleScriptString(
-    buildShellCommand(buildLocalTerminalShellArgs(args)),
-  );
-
-  if (args.appName === "Terminal") {
-    return [
-      "-e",
-      `tell application "Terminal" to do script "${command}"`,
-      "-e",
-      'tell application "Terminal" to activate',
-    ];
-  }
-
-  return [
-    "-e",
-    'tell application "iTerm" to create window with default profile',
-    "-e",
-    `tell application "iTerm" to tell current session of current window to write text "${command}"`,
-    "-e",
-    'tell application "iTerm" to activate',
-  ];
-}
-
-function buildMacTerminalRemoteSshOpenArgs(
-  args: BuildMacRemoteSshOpenArgs & { appName: "Terminal" | "iTerm" },
-): string[] {
-  const command = escapeAppleScriptString(
-    buildShellCommand(buildRemoteTerminalSshArgs(args)),
-  );
-
-  if (args.appName === "Terminal") {
-    return [
-      "-e",
-      `tell application "Terminal" to do script "${command}"`,
-      "-e",
-      'tell application "Terminal" to activate',
-    ];
-  }
-
-  return [
-    "-e",
-    'tell application "iTerm" to create window with default profile',
-    "-e",
-    `tell application "iTerm" to tell current session of current window to write text "${command}"`,
-    "-e",
-    'tell application "iTerm" to activate',
-  ];
-}
-
-function buildMacGhosttyLocalOpenArgs(
-  args: BuildMacTerminalOpenArgs,
-): string[] {
-  return [
-    "-na",
-    "Ghostty",
-    "--args",
-    "-e",
-    ...buildLocalTerminalShellArgs(args),
-  ];
-}
-
-function buildMacGhosttyRemoteSshOpenArgs(
-  args: BuildMacRemoteSshOpenArgs,
-): string[] {
-  return [
-    "-na",
-    "Ghostty",
-    "--args",
-    "-e",
-    ...buildRemoteTerminalSshArgs(args),
-  ];
-}
-
-const FULL_FILE_OPEN_CAPABILITIES: WorkspaceOpenTargetCapabilities = {
-  openDirectory: true,
-  openFile: true,
-  openFileAtColumn: true,
-  openFileAtLine: true,
-};
-
-const BASIC_FILE_OPEN_CAPABILITIES: WorkspaceOpenTargetCapabilities = {
-  openDirectory: true,
-  openFile: true,
-  openFileAtColumn: false,
-  openFileAtLine: false,
-};
-
-const FILE_MANAGER_OPEN_CAPABILITIES: WorkspaceOpenTargetCapabilities = {
-  openDirectory: true,
-  openFile: true,
-  openFileAtColumn: false,
-  openFileAtLine: false,
-};
-
-const TERMINAL_OPEN_CAPABILITIES: WorkspaceOpenTargetCapabilities = {
-  openDirectory: true,
-  openFile: true,
-  openFileAtColumn: true,
-  openFileAtLine: true,
-};
-
-const LINE_ONLY_FILE_OPEN_CAPABILITIES: WorkspaceOpenTargetCapabilities = {
-  openDirectory: true,
-  openFile: true,
-  openFileAtColumn: false,
-  openFileAtLine: true,
-};
-
-const LAUNCH_ADAPTERS: LaunchAdapter[] = [
-  {
-    capabilities: FULL_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "vscode" },
-    id: "vscode",
-    kind: "editor",
-    label: "VS Code",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Visual Studio Code",
-      bundleIds: ["com.microsoft.VSCode"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "code",
-        supportsColumn: true,
-        toArgs: (args) => ["-g", formatPathWithLineNumber(args)],
-      },
-      pathOpenCommand: {
-        executable: "code",
-        toArgs: (path) => [path],
-      },
-      remoteSshOpenCommand: {
-        capabilities: FULL_FILE_OPEN_CAPABILITIES,
-        executable: "code",
-        toArgs: (args) => [
-          "--remote",
-          `ssh-remote+${args.sshAuthority}`,
-          ...(args.lineNumber === null
-            ? [args.path]
-            : [
-                "-g",
-                formatPathWithLineNumber({
-                  lineNumber: args.lineNumber,
-                  columnNumber: args.columnNumber,
-                  path: args.path,
-                }),
-              ]),
-        ],
-      },
-    },
-  },
-  {
-    capabilities: FULL_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "vscode" },
-    id: "vscode-insiders",
-    kind: "editor",
-    label: "VS Code Insiders",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Visual Studio Code - Insiders",
-      bundleIds: ["com.microsoft.VSCodeInsiders"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "code-insiders",
-        supportsColumn: true,
-        toArgs: (args) => ["-g", formatPathWithLineNumber(args)],
-      },
-      pathOpenCommand: {
-        executable: "code-insiders",
-        toArgs: (path) => [path],
-      },
-      remoteSshOpenCommand: {
-        capabilities: FULL_FILE_OPEN_CAPABILITIES,
-        executable: "code-insiders",
-        toArgs: (args) => [
-          "--remote",
-          `ssh-remote+${args.sshAuthority}`,
-          ...(args.lineNumber === null
-            ? [args.path]
-            : [
-                "-g",
-                formatPathWithLineNumber({
-                  lineNumber: args.lineNumber,
-                  columnNumber: args.columnNumber,
-                  path: args.path,
-                }),
-              ]),
-        ],
-      },
-    },
-  },
-  {
-    capabilities: FULL_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "cursor" },
-    id: "cursor",
-    kind: "editor",
-    label: "Cursor",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Cursor",
-      // ToDesktop bundle IDs are generated; keep app-name path fallback below.
-      bundleIds: ["com.todesktop.230313mzl4w4u92"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "cursor",
-        supportsColumn: true,
-        toArgs: (args) => ["-g", formatPathWithLineNumber(args)],
-      },
-      pathOpenCommand: {
-        executable: "cursor",
-        toArgs: (path) => [path],
-      },
-      remoteSshOpenCommand: {
-        capabilities: FULL_FILE_OPEN_CAPABILITIES,
-        executable: "cursor",
-        toArgs: (args) => [
-          "--remote",
-          `ssh-remote+${args.sshAuthority}`,
-          ...(args.lineNumber === null
-            ? [args.path]
-            : [
-                "-g",
-                formatPathWithLineNumber({
-                  lineNumber: args.lineNumber,
-                  columnNumber: args.columnNumber,
-                  path: args.path,
-                }),
-              ]),
-        ],
-      },
-    },
-  },
-  {
-    capabilities: FULL_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "sublime-text" },
-    id: "sublime-text",
-    kind: "editor",
-    label: "Sublime Text",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Sublime Text",
-      bundleIds: ["com.sublimetext.4", "com.sublimetext.3"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "subl",
-        supportsColumn: true,
-        toArgs: (args) => [formatPathWithLineNumber(args)],
-      },
-    },
-  },
-  {
-    capabilities: FULL_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "zed" },
-    id: "zed",
-    kind: "editor",
-    label: "Zed",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Zed",
-      bundleIds: ["dev.zed.Zed"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "zed",
-        supportsColumn: true,
-        toArgs: (args) => [formatPathWithLineNumber(args)],
-      },
-      pathOpenCommand: {
-        executable: "zed",
-        toArgs: (path) => [path],
-      },
-      remoteSshOpenCommand: {
-        capabilities: FULL_FILE_OPEN_CAPABILITIES,
-        executable: "zed",
-        toArgs: (args) => [formatZedRemoteSshUri(args)],
-      },
-    },
-  },
-  {
-    capabilities: FULL_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "windsurf" },
-    id: "windsurf",
-    kind: "editor",
-    label: "Windsurf",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Windsurf",
-      bundleIds: ["com.exafunction.windsurf"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "windsurf",
-        supportsColumn: true,
-        toArgs: (args) => ["-g", formatPathWithLineNumber(args)],
-      },
-      pathOpenCommand: {
-        executable: "windsurf",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "bbedit",
-    kind: "editor",
-    label: "BBEdit",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "BBEdit",
-      bundleIds: ["com.barebones.bbedit"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "bbedit",
-        supportsColumn: false,
-        toArgs: (args) => [`+${args.lineNumber}`, args.path],
-      },
-      pathOpenCommand: {
-        executable: "bbedit",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "textmate",
-    kind: "editor",
-    label: "TextMate",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "TextMate",
-      bundleIds: ["com.macromates.TextMate"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "mate",
-        supportsColumn: false,
-        toArgs: (args) => ["--line", String(args.lineNumber), args.path],
-      },
-      pathOpenCommand: {
-        executable: "mate",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: FULL_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "emacs",
-    kind: "editor",
-    label: "Emacs",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Emacs",
-      bundleIds: ["org.gnu.Emacs"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "emacsclient",
-        supportsColumn: true,
-        toArgs: (args) => [
-          args.columnNumber === null
-            ? `+${args.lineNumber}`
-            : `+${args.lineNumber}:${args.columnNumber}`,
-          args.path,
-        ],
-      },
-      pathOpenCommand: {
-        executable: "emacsclient",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "intellij-idea",
-    kind: "editor",
-    label: "IntelliJ IDEA",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "IntelliJ IDEA",
-      bundleIds: ["com.jetbrains.intellij", "com.jetbrains.intellij.ce"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "idea",
-        supportsColumn: false,
-        toArgs: formatJetBrainsLineOpenArgs,
-      },
-      pathOpenCommand: {
-        executable: "idea",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "pycharm",
-    kind: "editor",
-    label: "PyCharm",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "PyCharm",
-      bundleIds: ["com.jetbrains.pycharm", "com.jetbrains.pycharm.ce"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "pycharm",
-        supportsColumn: false,
-        toArgs: formatJetBrainsLineOpenArgs,
-      },
-      pathOpenCommand: {
-        executable: "pycharm",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "webstorm",
-    kind: "editor",
-    label: "WebStorm",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "WebStorm",
-      bundleIds: ["com.jetbrains.WebStorm"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "webstorm",
-        supportsColumn: false,
-        toArgs: formatJetBrainsLineOpenArgs,
-      },
-      pathOpenCommand: {
-        executable: "webstorm",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "goland",
-    kind: "editor",
-    label: "GoLand",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "GoLand",
-      bundleIds: ["com.jetbrains.goland"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "goland",
-        supportsColumn: false,
-        toArgs: formatJetBrainsLineOpenArgs,
-      },
-      pathOpenCommand: {
-        executable: "goland",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "rider",
-    kind: "editor",
-    label: "Rider",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Rider",
-      bundleIds: ["com.jetbrains.rider"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "rider",
-        supportsColumn: false,
-        toArgs: formatJetBrainsLineOpenArgs,
-      },
-      pathOpenCommand: {
-        executable: "rider",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "rustrover",
-    kind: "editor",
-    label: "RustRover",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "RustRover",
-      bundleIds: ["com.jetbrains.rustrover"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "rustrover",
-        supportsColumn: false,
-        toArgs: formatJetBrainsLineOpenArgs,
-      },
-      pathOpenCommand: {
-        executable: "rustrover",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "phpstorm",
-    kind: "editor",
-    label: "PhpStorm",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "PhpStorm",
-      bundleIds: ["com.jetbrains.PhpStorm"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "phpstorm",
-        supportsColumn: false,
-        toArgs: formatJetBrainsLineOpenArgs,
-      },
-      pathOpenCommand: {
-        executable: "phpstorm",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "app" },
-    id: "android-studio",
-    kind: "editor",
-    label: "Android Studio",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Android Studio",
-      bundleIds: ["com.google.android.studio"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "studio",
-        supportsColumn: false,
-        toArgs: formatJetBrainsLineOpenArgs,
-      },
-      pathOpenCommand: {
-        executable: "studio",
-        toArgs: (path) => [path],
-      },
-    },
-  },
-  {
-    capabilities: BASIC_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "antigravity" },
-    id: "antigravity",
-    kind: "editor",
-    label: "Antigravity",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Antigravity",
-      bundleIds: ["com.google.antigravity", "com.googlelabs.antigravity"],
-      builtIn: false,
-    },
-  },
-  {
-    capabilities: LINE_ONLY_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "xcode" },
-    id: "xcode",
-    kind: "editor",
-    label: "Xcode",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Xcode",
-      bundleIds: ["com.apple.dt.Xcode"],
-      builtIn: false,
-      lineOpenCommand: {
-        executable: "xed",
-        supportsColumn: false,
-        toArgs: (args) => ["-l", String(args.lineNumber), args.path],
-      },
-    },
-  },
-  {
-    capabilities: FILE_MANAGER_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "finder" },
-    id: "finder",
-    kind: "file-manager",
-    label: "Finder",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "application",
-      appName: "Finder",
-      bundleIds: ["com.apple.finder"],
-      builtIn: true,
-      fileOpenCommand: {
-        executable: "open",
-        toArgs: (path) => ["-R", path],
-      },
-    },
-  },
-  {
-    capabilities: TERMINAL_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "terminal" },
-    id: "terminal",
-    kind: "terminal",
-    label: "Terminal",
-    fileOpenBehavior: "containing-directory",
-    macos: {
-      openMode: "application",
-      appName: "Terminal",
-      bundleIds: ["com.apple.Terminal"],
-      builtIn: true,
-      localTerminalOpenCommand: {
-        executable: "osascript",
-        toArgs: (args) =>
-          buildMacTerminalLocalOpenArgs({
-            appName: "Terminal",
-            columnNumber: args.columnNumber,
-            lineNumber: args.lineNumber,
-            path: args.path,
-          }),
-      },
-      remoteSshOpenCommand: {
-        capabilities: TERMINAL_OPEN_CAPABILITIES,
-        executable: "osascript",
-        requiredExecutables: ["ssh"],
-        toArgs: (args) =>
-          buildMacTerminalRemoteSshOpenArgs({
-            appName: "Terminal",
-            columnNumber: args.columnNumber,
-            lineNumber: args.lineNumber,
-            path: args.path,
-            sshAuthority: args.sshAuthority,
-          }),
-      },
-    },
-  },
-  {
-    capabilities: TERMINAL_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "iterm2" },
-    id: "iterm2",
-    kind: "terminal",
-    label: "iTerm2",
-    fileOpenBehavior: "containing-directory",
-    macos: {
-      openMode: "application",
-      appName: "iTerm",
-      bundleIds: ["com.googlecode.iterm2"],
-      builtIn: false,
-      localTerminalOpenCommand: {
-        executable: "osascript",
-        toArgs: (args) =>
-          buildMacTerminalLocalOpenArgs({
-            appName: "iTerm",
-            columnNumber: args.columnNumber,
-            lineNumber: args.lineNumber,
-            path: args.path,
-          }),
-      },
-      remoteSshOpenCommand: {
-        capabilities: TERMINAL_OPEN_CAPABILITIES,
-        executable: "osascript",
-        requiredExecutables: ["ssh"],
-        toArgs: (args) =>
-          buildMacTerminalRemoteSshOpenArgs({
-            appName: "iTerm",
-            columnNumber: args.columnNumber,
-            lineNumber: args.lineNumber,
-            path: args.path,
-            sshAuthority: args.sshAuthority,
-          }),
-      },
-    },
-  },
-  {
-    capabilities: TERMINAL_OPEN_CAPABILITIES,
-    icon: { kind: "builtin", name: "ghostty" },
-    id: "ghostty",
-    kind: "terminal",
-    label: "Ghostty",
-    fileOpenBehavior: "containing-directory",
-    macos: {
-      openMode: "application",
-      appName: "Ghostty",
-      bundleIds: ["com.mitchellh.ghostty"],
-      builtIn: false,
-      localTerminalOpenCommand: {
-        executable: "open",
-        toArgs: (args) =>
-          buildMacGhosttyLocalOpenArgs({
-            columnNumber: args.columnNumber,
-            lineNumber: args.lineNumber,
-            path: args.path,
-          }),
-      },
-      remoteSshOpenCommand: {
-        capabilities: TERMINAL_OPEN_CAPABILITIES,
-        executable: "open",
-        requiredExecutables: ["ssh"],
-        toArgs: (args) =>
-          buildMacGhosttyRemoteSshOpenArgs({
-            columnNumber: args.columnNumber,
-            lineNumber: args.lineNumber,
-            path: args.path,
-            sshAuthority: args.sshAuthority,
-          }),
-      },
-    },
-  },
-  {
-    capabilities: BASIC_FILE_OPEN_CAPABILITIES,
-    icon: { kind: "symbol", name: "default-app" },
-    id: "default-app",
-    kind: "default-app",
-    label: "Default App",
-    fileOpenBehavior: "direct",
-    macos: {
-      openMode: "default-app",
-    },
-  },
-];
 
 async function toWorkspaceOpenTarget(
   definition: LaunchAdapter,
@@ -1817,11 +855,6 @@ function requireLaunchAdapter(targetId: WorkspaceOpenTargetId): LaunchAdapter {
     });
   }
   return definition;
-}
-
-interface ExistingPath {
-  path: string;
-  type: "directory" | "file";
 }
 
 async function requireOpenablePath(targetPath: string): Promise<ExistingPath> {
